# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NEXORA Platform APIs",
    "description": "Enterprise-grade REST and WebSockets platform API spec supporting the NEXORA Predictive Crowd Intelligence platform.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Primary Backend Microserver"
    },
    {
      "url": "http://localhost:8001",
      "description": "Real-time Map Socket Server"
    }
  ],
  "security": [
    {
      "OAuth2Bearer": []
    }
  ],
  "paths": {
...
    "/xai/explain": {
      "post": {
        "summary": "SHAP Explainability Analytics",
        "description": "Computes impact index factors contributing to risk assessments.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SHAP features and raw text reasons compiled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XAIExplanationOutput"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OAuth2Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PredictionInput": {
        "type": "object",
        "properties": {
          "total_headcount": { "type": "integer" },
          "occupancy_pct": { "type": "number" },
          "density_value": { "type": "number" },
          "queue_length": { "type": "integer" },
          "flow_rate_delta": { "type": "number" }
        },
        "required": ["total_headcount", "occupancy_pct", "density_value", "queue_length", "flow_rate_delta"]
      },
...
      "XAIExplanationOutput": {
        "type": "object",
        "properties": {
          "shap_base_value": { "type": "number" },
          "prediction_probability": { "type": "number" },
          "contributions": {
            "type": "object",
            "additionalProperties": { "type": "number" }
          },
          "explanation_text": { "type": "string" }
        }
      },
...
    }
  }
}